container-view extraction
